+++ config/packages/twig.php
@@ -16,6 +16,7 @@ return static function (ContainerConfigurator $containerConfigurator): void {
         'app_version' => '%env(APP_VERSION)%',
         'app_env' => '%env(APP_ENV)%',
         'bugsnag_api_key' => '%env(BUGSNAG_API_KEY)%',
+        'gtm_container_id' => '%env(GTM_CONTAINER_ID)%',
       ],
       'form_themes' => [
         '@SonataForm/Form/datepicker.html.twig',