+++ src/Application/Controller/Base/TagExtensionController.php
@@ -6,7 +6,7 @@ namespace App\Application\Controller\Base;
 
 use Symfony\Bundle\FrameworkBundle\Controller\AbstractController;
 use Symfony\Component\HttpFoundation\Response;
-use Symfony\Component\Routing\Annotation\Route;
+use Symfony\Component\Routing\Attribute\Route;
 
 class TagExtensionController extends AbstractController
 {