+++ src/DB/Entity/Project/ProgramDownloads.php
@@ -11,8 +11,8 @@ use Doctrine\ORM\Mapping as ORM;
 #[ORM\Entity]
 class ProgramDownloads
 {
-  final public const TYPE_PROJECT = 'project';
-  final public const TYPE_APK = 'apk';
+  final public const string TYPE_PROJECT = 'project';
+  final public const string TYPE_APK = 'apk';
 
   #[ORM\Column(name: 'id', type: 'integer')]
   #[ORM\Id]